+++ src/components/screen/Ad/simpleForm/InitialValues.js
@@ -1,0 +1,11 @@
+const initialValues = {
+    title: 'Titulo',
+    price: 1,
+    describe: 'description',
+    visibleFrom: 'visibleFrom',
+    visibleTo: 'visible',
+    image: '',
+    estabilishment_id: '4ROchfT9WUZ4gQeaE2mH'
+}
+
+export default initialValues
