--- src/components/screen/Anuncio/simpleForm/InitialValues.js
@@ -1,10 +1,0 @@
-const initialValues = {
-    title: '',
-    price: '',
-    describe: '',
-    visibleFrom: '',
-    visibleUntil: '',
-    image: ''
-}
-
-export default initialValues
