--- src/components/screen/Anuncio/simpleForm/Validation.js
@@ -1,6 +1,0 @@
-import * as yup from 'yup';
-
-const validationSchema = yup.object().shape({
-    title:: yup.string()
-    .
-})
