+++ src/services/api.js
@@ -1,7 +1,8 @@
 import axios from 'axios';
 
 const api = axios.create({
-  baseURL: 'https://petmagnet-api.herokuapp.com'
+  //baseURL: 'https://petmagnet-api.herokuapp.com'
+  baseURL: 'https://us-central1-petsmagnet.cloudfunctions.net'
 });
 
 export default api;
